News_research_tool/requirements.txt
langchain==0.3.0
langchain_groq==0.2.0
numpy==1.23.2
python-dotenv==1.0.1
sentence_transformers==3.1.0
streamlit==1.38.0